// types/letter/letterRequest.ts
export default interface LetterRequest{
    no_letter: number
    date_letter: string
    date_entry: string
    letter_type: string
    regarding: string
    attachment: string
    letter_form: string
    status: string
    characteristic: string
    attachment_file_url: string
    letter_out : boolean
}